+++ src/joLinearFunctionConstructors/joCurvelet2DnoFFT.jl
@@ -3,7 +3,7 @@
 ## helper module
 module joCurvelet2DnoFFT_etc
     using JOLI: jo_convert, jo_complex_eltype
-    function apply_fdct2DnoFFTwrap_real(v::AbstractVector,n1::Integer,n2::Integer,m::Integer,rdt::DataType,nbs::Integer,nbac::Integer,actl::Integer,rctl::Integer,zfin::Integer)
+    function apply_fdct2DnoFFTwrap_real(v::AbstractVector,n1::Integer,n2::Integer,m::Int128,rdt::DataType,nbs::Integer,nbac::Integer,actl::Integer,rctl::Integer,zfin::Integer)
         C=Vector{Cdouble}(m)
         X=jo_convert(Complex{Cdouble},v,false)
         ccall((:jl_fdct_wrapping_real_nofft,:libdfdct_wrapping),Void,
@@ -12,7 +12,7 @@ module joCurvelet2DnoFFT_etc
         C=jo_convert(rdt,C,false)
         return C
     end
-    function apply_ifdct2DnoFFTwrap_real(v::AbstractVector,n1::Integer,n2::Integer,m::Integer,rdt::DataType,nbs::Integer,nbac::Integer,actl::Integer,rctl::Integer,zfin::Integer)
+    function apply_ifdct2DnoFFTwrap_real(v::AbstractVector,n1::Integer,n2::Integer,m::Int128,rdt::DataType,nbs::Integer,nbac::Integer,actl::Integer,rctl::Integer,zfin::Integer)
         X=Vector{Complex{Cdouble}}(n1*n2)
         eltype(v)<:Real || throw(joLinearFunctionException("joCurvelt2DnoFFT: input vector must be real for real transform"))
         C=jo_convert(Cdouble,v,false)
@@ -22,7 +22,7 @@ module joCurvelet2DnoFFT_etc
         X=jo_convert(rdt,X,false)
         return X
     end
-    function apply_fdct2DnoFFTwrap_cplx(v::AbstractVector,n1::Integer,n2::Integer,m::Integer,rdt::DataType,nbs::Integer,nbac::Integer,actl::Integer,rctl::Integer,zfin::Integer)
+    function apply_fdct2DnoFFTwrap_cplx(v::AbstractVector,n1::Integer,n2::Integer,m::Int128,rdt::DataType,nbs::Integer,nbac::Integer,actl::Integer,rctl::Integer,zfin::Integer)
         C=Vector{Complex{Cdouble}}(m)
         X=jo_convert(Complex{Cdouble},v,false)
         ccall((:jl_fdct_wrapping_cpx_nofft,:libdfdct_wrapping),Void,
@@ -32,7 +32,7 @@ module joCurvelet2DnoFFT_etc
         C=jo_convert(rdt,C,false)
         return C
     end
-    function apply_ifdct2DnoFFTwrap_cplx(v::AbstractVector,n1::Integer,n2::Integer,m::Integer,rdt::DataType,nbs::Integer,nbac::Integer,actl::Integer,rctl::Integer,zfin::Integer)
+    function apply_ifdct2DnoFFTwrap_cplx(v::AbstractVector,n1::Integer,n2::Integer,m::Int128,rdt::DataType,nbs::Integer,nbac::Integer,actl::Integer,rctl::Integer,zfin::Integer)
         X=Vector{Complex{Cdouble}}(n1*n2)
         C=jo_convert(Complex{Cdouble},v,false)
         ccall((:jl_ifdct_wrapping_cpx_nofft,:libdfdct_wrapping),Void,
